Moved mod cli to a its own folder and file
