+++ src/lib.rs
@@ -1,1 +1,2 @@
 pub mod files;
+pub mod cli;
